cifssrv: remove dceprc and winreg codes in cifssrv
since moving dcerpc and winreg to userspace, these codes have been not used.
So this patch remove dceprc and winreg codes in cifssrv.
