+++ smb2pdu.c
@@ -22,7 +22,6 @@
 #include "glob.h"
 #include "export.h"
 #include "smb2pdu.h"
-#include "dcerpc.h"
 #include "smbfsctl.h"
 #include "oplock.h"
 
@@ -1190,10 +1189,8 @@ int smb2_sess_setup(struct smb_work *smb_work)
 		if (rc < 0)
 			goto out_err;
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 		init_waitqueue_head(&sess->pipe_q);
 		sess->ev_state = NETLINK_REQ_INIT;
-#endif
 	} else {
 		struct smb2_hdr *req_hdr = (struct smb2_hdr *)smb_work->buf;
 
@@ -1787,9 +1784,7 @@ static int create_smb2_pipe(struct smb_work *smb_work)
 	struct smb2_create_rsp *rsp;
 	struct smb2_create_req *req;
 	int id;
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	int err;
-#endif
 	unsigned int pipe_type;
 	char *name;
 
@@ -1819,12 +1814,10 @@ static int create_smb2_pipe(struct smb_work *smb_work)
 		return id;
 	}
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	err = cifssrv_sendmsg(smb_work->sess,
 			CIFSSRV_KEVENT_CREATE_PIPE, pipe_type, 0, NULL, 0);
 	if (err)
 		cifssrv_err("failed to send event, err %d\n", err);
-#endif
 
 	rsp->StructureSize = cpu_to_le16(89);
 	rsp->OplockLevel = SMB2_OPLOCK_LEVEL_NONE;
@@ -3311,7 +3304,6 @@ static int smb2_close_pipe(struct smb_work *smb_work)
 	rsp->Attributes = 0;
 	inc_rfc1001_len(rsp, 60);
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	if (!rc) {
 		rc = cifssrv_sendmsg(smb_work->sess,
 				CIFSSRV_KEVENT_DESTROY_PIPE,
@@ -3319,7 +3311,7 @@ static int smb2_close_pipe(struct smb_work *smb_work)
 		if (rc)
 			cifssrv_err("failed to send event, err %d\n", rc);
 	}
-#endif
+
 	rc = close_pipe_id(smb_work->sess, pipe_desc->pipe_type);
 	if (rc < 0) {
 		rsp->hdr.Status = NT_STATUS_INVALID_HANDLE;
@@ -4892,9 +4884,7 @@ int smb2_read_pipe(struct smb_work *smb_work)
 	struct smb2_read_req *req;
 	struct smb2_read_rsp *rsp;
 	unsigned int read_len;
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	struct cifssrv_uevent *ev;
-#endif
 	struct cifssrv_pipe *pipe_desc;
 	req = (struct smb2_read_req *)smb_work->buf;
 	rsp = (struct smb2_read_rsp *)smb_work->rsp_buf;
@@ -4911,7 +4901,6 @@ int smb2_read_pipe(struct smb_work *smb_work)
 		return ret;
 	}
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	ret = cifssrv_sendmsg(smb_work->sess, CIFSSRV_KEVENT_READ_PIPE,
 			pipe_desc->pipe_type, 0, NULL, read_len);
 	if (ret)
@@ -4929,16 +4918,6 @@ int smb2_read_pipe(struct smb_work *smb_work)
 		memcpy(data_buf, pipe_desc->rsp_buf, nbytes);
 		smb_work->sess->ev_state = NETLINK_REQ_COMPLETED;
 	}
-#else
-	nbytes = process_rpc_rsp(smb_work->sess, data_buf, read_len,
-			pipe_desc->pipe_type);
-	if (nbytes <= 0) {
-		cifssrv_err("Pipe data not present\n");
-		rsp->hdr.Status = NT_STATUS_UNEXPECTED_IO_ERROR;
-		smb2_set_err_rsp(smb_work);
-		return -EINVAL;
-	}
-#endif
 
 	rsp->StructureSize = cpu_to_le16(17);
 	rsp->DataOffset = 80;
@@ -5075,9 +5054,7 @@ static int smb2_write_pipe(struct smb_work *smb_work)
 	int err = 0, ret = 0;
 	char *data_buf;
 	size_t length;
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	struct cifssrv_uevent *ev;
-#endif
 	struct cifssrv_pipe *pipe_desc;
 
 	req = (struct smb2_write_req *)smb_work->buf;
@@ -5112,7 +5089,6 @@ static int smb2_write_pipe(struct smb_work *smb_work)
 				le16_to_cpu(req->DataOffset));
 	}
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	ret = cifssrv_sendmsg(smb_work->sess, CIFSSRV_KEVENT_WRITE_PIPE,
 			pipe_desc->pipe_type, length, data_buf, 0);
 	if (ret)
@@ -5133,18 +5109,6 @@ static int smb2_write_pipe(struct smb_work *smb_work)
 		length = ev->u.w_pipe_rsp.write_count;
 		smb_work->sess->ev_state = NETLINK_REQ_COMPLETED;
 	}
-#else
-	ret = process_rpc(smb_work->sess, data_buf, pipe_desc->pipe_type);
-	if (ret == -EOPNOTSUPP) {
-		rsp->hdr.Status = NT_STATUS_NOT_SUPPORTED;
-		smb2_set_err_rsp(smb_work);
-		return ret;
-	} else if (ret) {
-		rsp->hdr.Status = NT_STATUS_INVALID_HANDLE;
-		smb2_set_err_rsp(smb_work);
-		return ret;
-	}
-#endif
 
 	rsp->StructureSize = cpu_to_le16(17);
 	rsp->DataOffset = 0;
@@ -5810,9 +5774,7 @@ int smb2_ioctl(struct smb_work *smb_work)
 	uint64_t id = -1;
 	int ret = 0;
 	struct tcp_server_info *server = smb_work->server;
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	struct cifssrv_uevent *ev;
-#endif
 	struct cifssrv_pipe *pipe_desc;
 
 	req = (struct smb2_ioctl_req *)smb_work->buf;
@@ -5842,9 +5804,7 @@ int smb2_ioctl(struct smb_work *smb_work)
 
 	cnt_code = le32_to_cpu(req->CntCode);
 	out_buf_len = le32_to_cpu(req->maxoutputresp);
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 	out_buf_len = min(NETLINK_CIFSSRV_MAX_PAYLOAD, out_buf_len);
-#endif
 	data_buf = (char *)&req->Buffer[0];
 
 	switch (cnt_code) {
@@ -5884,7 +5844,6 @@ int smb2_ioctl(struct smb_work *smb_work)
 			goto out;
 		}
 
-#ifdef CONFIG_CIFSSRV_NETLINK_INTERFACE
 		ret = cifssrv_sendmsg(smb_work->sess, CIFSSRV_KEVENT_IOCTL_PIPE,
 				pipe_desc->pipe_type,
 				le32_to_cpu(req->inputcount), data_buf,
@@ -5919,31 +5878,6 @@ int smb2_ioctl(struct smb_work *smb_work)
 					nbytes);
 			smb_work->sess->ev_state = NETLINK_REQ_COMPLETED;
 		}
-#else
-		ret = process_rpc(smb_work->sess, data_buf,
-			pipe_desc->pipe_type);
-		if (ret == -EOPNOTSUPP) {
-			rsp->hdr.Status =
-				NT_STATUS_NOT_SUPPORTED;
-			goto out;
-		} else if (ret) {
-			rsp->hdr.Status =
-				NT_STATUS_INVALID_PARAMETER;
-			goto out;
-		}
-
-		nbytes = process_rpc_rsp(smb_work->sess, (char *)rsp->Buffer,
-				out_buf_len, pipe_desc->pipe_type);
-		if (nbytes > out_buf_len) {
-			rsp->hdr.Status =
-				NT_STATUS_BUFFER_OVERFLOW;
-			nbytes = out_buf_len;
-		} else if (nbytes < 0) {
-			rsp->hdr.Status =
-				NT_STATUS_INVALID_PARAMETER;
-			goto out;
-		}
-#endif
 
 		break;
 	case FSCTL_VALIDATE_NEGOTIATE_INFO:
